pyproject.toml
[tool.poetry]
name = "document-intelligence"
version = "1.0.0"
description = "Production-grade document intelligence system for extracting structured data from PDFs and images"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
asyncpg = "^0.29.0"
alembic = "^1.13.1"
redis = "^5.0.1"
arq = "^0.25.0"
anthropic = "^0.40.0"
openai = "^1.51.0"
pypdf = "^3.17.4"
pdf2image = "^1.17.0"
pdfplumber = "^0.11.0"
pillow = "^10.2.0"
opencv-python = "^4.9.0"
pytesseract = "^0.3.10"
boto3 = "^1.34.34"
python-multipart = "^0.0.6"
aiofiles = "^23.2.1"
httpx = "^0.26.0"
prometheus-client = "^0.19.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
tenacity = "^8.2.3"
structlog = "^24.1.0"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
black = "^24.1.1"
ruff = "^0.1.14"
mypy = "^1.8.0"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=html --cov-report=term-missing"
